+++ src/Editor/Lyrics/LyricPreview/EditableTextInput.tsx
