+++ src/Editor/Lyrics/LyricPreview/LyricsTextView.tsx
