+++ src/Editor/Lyrics/LyricPreview/ResizableText.tsx
